uni_profile_management_service/pyproject.toml
[tool.poetry]
name = "uni-profile-management-service"
version = "0.1.0"
description = ""
authors = ["Zain <ziynali938@gmail.com>"]
package-mode=false 
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.2.3"
pydantic = "^2.6.4"
emails = "^0.6"

gunicorn = "^21.2.0"
jinja2 = "^3.1.3"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"


# Pin bcrypt until passlib supports the latest
bcrypt = "4.0.1"
pydantic-settings = "^2.2.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.42.0"}
aiokafka = "^0.10.0"
psycopg2 = "^2.9.9"
protobuf = "^5.27.3"
confluent-kafka = "^2.5.3"
pyjwt = "^2.9.0"
pytz = "^2024.2"
fastapi = "^0.115.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
